refactor(kubeproxy): move package
